tests: move git() method to util lib
So we can share this method among more tests
